+++ src/cookie_consent/admin.py
@@ -1,0 +1,14 @@
+from django.contrib.admin import ModelAdmin, register
+from solo.admin import SingletonModelAdmin
+
+from cookie_consent.models import BannerConfig, CookieGroup
+
+
+@register(BannerConfig)
+class BannerConfigAdmin(SingletonModelAdmin):
+    pass
+
+
+@register(CookieGroup)
+class CookieGroupAdmin(ModelAdmin):
+    readonly_fields = ("version",)