+++ src/cookie_consent/models.py
@@ -1,0 +1,27 @@
+from django.db import models
+from solo.models import SingletonModel
+
+
+class BannerConfig(SingletonModel):  # type: ignore[django-manager-missing] # https://github.com/typeddjango/django-stubs/issues/1023
+    title = models.CharField(max_length=100, unique=True)
+    text = models.CharField(max_length=200)
+    show_banner = models.BooleanField(default=False)
+
+    def __str__(self) -> str:
+        return "Cookie consent banner"
+
+
+class CookieGroup(models.Model):  # type: ignore[django-manager-missing] # https://github.com/typeddjango/django-stubs/issues/1023
+    cookie_id = models.CharField(max_length=50, primary_key=True)
+    name = models.CharField(max_length=50)
+    description = models.TextField()
+    is_required = models.BooleanField()
+    version = models.PositiveBigIntegerField(default=1, editable=False)
+
+    def __str__(self) -> str:
+        return self.name
+
+    def save(self, *args, **kwargs):
+        if self.pk:
+            self.version += 1
+        super().save(*args, **kwargs)